backend/requirements.txt
boto3
colorama
defog==0.67.2
fastapi
google-auth
httpx
hypercorn
numpy
openai
pandas==2.2.3
pgvector==0.3.6
pyyaml
redis
requests
resend
slack_sdk
sqlalchemy
sqlalchemy[asyncio]
sqlalchemy-utils
sqlglot
sqlparse

# sync db connectors
databricks-sql-connector
google-cloud-bigquery
mysql-connector-python
psycopg2-binary
pyodbc
redshift_connector
snowflake-connector-python


# async db connectors
asyncpg
aiomysql
aiosqlite
aioodbc

# celery
celery[redis]==5.4.0

# tests
pytest
pytest-asyncio

# llm providers
openai==1.61.0
anthropic==0.43.1
google-genai==0.5.0